requirements.txt
# Core API
fastapi==0.128.0
uvicorn==0.40.0
starlette==0.50.0

# Database
SQLAlchemy==2.0.46
psycopg2-binary==2.9.11

# Auth & Security (PINNED to avoid bcrypt crash)
python-jose[cryptography]==3.5.0
passlib==1.7.4
bcrypt==3.2.2

# File upload & parsing
python-multipart==0.0.22
pandas==2.3.3
openpyxl>=3.1.2

# Environment variables
python-dotenv==1.2.1

# Validation
pydantic[email]==2.12.5

# HTTP / API
httpx==0.28.1
requests==2.32.5

# AI (only what you use)
langchain==1.2.8
langchain-core==1.2.8
langchain-openai==1.1.7
langchain-groq==1.1.2

# Optional (PDF reports)
reportlab==4.2.0
